zshrc: use standard paths to select the icon
Reduce the number of calls to awk to one.

Use the standard paths `/etc/os-release` with `/usr/lib/os-release` as a
fallback. Virtually all distros are compliant with this.

ID is always lowercase, no need to check.

Some distros (RHEL) have their ID in quotes.

Tested that all distros listed have the files and matching ID.